project/backend/requirements.txt
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
drf-yasg==1.21.7
Pillow==11.0.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0
isort==5.12.0
gunicorn==21.2.0
drf-nested-routers==0.95.0
django-filter==23.3